from-zero-to-hero/requirements-260127.txt
azure-ai-agentserver-agentframework==1.0.0b9 # inconpatible: with 260127
agent-framework==1.0.0b260127
agent-framework-a2a==1.0.0b260127
agent-framework-ag-ui==1.0.0b260127
agent-framework-anthropic==1.0.0b260127
agent-framework-azure-ai==1.0.0b260127
agent-framework-azure-ai-search==1.0.0b260127
agent-framework-azurefunctions==1.0.0b260127
agent-framework-chatkit==1.0.0b260127
agent-framework-copilotstudio==1.0.0b260127
agent-framework-core==1.0.0b260127
agent-framework-declarative==1.0.0b260127
agent-framework-devui==1.0.0b260127
agent-framework-lab==1.0.0b251024
agent-framework-mem0==1.0.0b260127
agent-framework-ollama==1.0.0b260127
agent-framework-purview==1.0.0b260127
agent-framework-redis==1.0.0b260127